pyproject.toml
[project]
name = "ai-healer-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.125.0",
    "langchain>=1.2.0",
    "langchain-ollama>=1.0.1",
    "playwright>=1.57.0",
    "pydantic>=2.12.5",
    "pytest-playwright>=0.7.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]

[tool.setuptools.packages.find]
exclude = ["artifacts*"]
